+++ lib/ts/recipe/thirdpartypasswordless/index.ts
@@ -316,7 +316,14 @@ export default class Wrapper {
         });
     }
 
-    static verifyCode(
+    /**
+     * This function will only verify the code (not consume it), and:
+     * NOT create a new user if it doesn't exist
+     * NOT verify the user email if it exists
+     * NOT do any linking
+     * NOT delete the code
+     */
+    static checkCode(
         input:
             | {
                   preAuthSessionId: string;
@@ -348,9 +355,8 @@ export default class Wrapper {
           }
         | { status: "RESTART_FLOW_ERROR" }
     > {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.verifyCode({
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.checkCode({
             ...input,
-            deleteCode: true,
             userContext: getUserContext(input.userContext),
         });
     }
@@ -514,7 +520,7 @@ export let revokeAllCodes = Wrapper.revokeAllCodes;
 
 export let revokeCode = Wrapper.revokeCode;
 
-export let verifyCode = Wrapper.verifyCode;
+export let checkCode = Wrapper.checkCode;
 
 export let createMagicLink = Wrapper.createMagicLink;