+++ lib/ts/recipe/thirdpartypasswordless/recipeImplementation/index.ts
@@ -26,8 +26,8 @@ export default function getRecipeInterface(
         consumeCode: async function (input) {
             return originalPasswordlessImplementation.consumeCode.bind(DerivedPwdless(this))(input);
         },
-        verifyCode: async function (input) {
-            return originalPasswordlessImplementation.verifyCode.bind(DerivedPwdless(this))(input);
+        checkCode: async function (input) {
+            return originalPasswordlessImplementation.checkCode.bind(DerivedPwdless(this))(input);
         },
         createCode: async function (input) {
             return originalPasswordlessImplementation.createCode.bind(DerivedPwdless(this))(input);