+++ lib/ts/recipe/thirdpartypasswordless/recipeImplementation/passwordlessRecipeImplementation.ts
@@ -6,8 +6,8 @@ export default function getRecipeInterface(recipeInterface: ThirdPartyPasswordle
         consumeCode: async function (input) {
             return await recipeInterface.consumeCode(input);
         },
-        verifyCode: async function (input) {
-            return await recipeInterface.verifyCode(input);
+        checkCode: async function (input) {
+            return await recipeInterface.checkCode(input);
         },
         createCode: async function (input) {
             return await recipeInterface.createCode(input);