+++ lib/ts/recipe/thirdpartypasswordless/types.ts
@@ -294,20 +294,18 @@ export type RecipeInterface = {
           }
     >;
 
-    verifyCode: (
+    checkCode: (
         input:
             | {
                   userInputCode: string;
                   deviceId: string;
                   preAuthSessionId: string;
-                  deleteCode: boolean;
                   tenantId: string;
                   userContext: UserContext;
               }
             | {
                   linkCode: string;
                   preAuthSessionId: string;
-                  deleteCode: boolean;
                   tenantId: string;
                   userContext: UserContext;
               }