+++ test/thirdpartypasswordless/smsDelivery.test.js
@@ -75,7 +75,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         let codeLifetime = undefined;
         let urlWithLinkCode = undefined;
         let userInputCode = undefined;
-        let isFirstFactor = undefined;
         let apiKey = "randomKey";
 
         nock("https://api.supertokens.com")
@@ -86,7 +85,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
                 codeLifetime = body.smsInput.codeLifetime;
                 urlWithLinkCode = body.smsInput.urlWithLinkCode;
                 userInputCode = body.smsInput.userInputCode;
-                isFirstFactor = body.smsInput.isFirstFactor;
                 apiKey = body.apiKey;
                 return {};
             });
@@ -109,7 +107,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         assert.notStrictEqual(userInputCode, undefined);
         assert.notStrictEqual(codeLifetime, undefined);
         assert.strictEqual(apiKey, undefined);
-        assert(isFirstFactor);
         assert(codeLifetime > 0);
     });
 
@@ -403,7 +400,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         let codeLifetime = undefined;
         let urlWithLinkCode = undefined;
         let userInputCode = undefined;
-        let isFirstFactor = undefined;
 
         nock("https://api.supertokens.com")
             .post("/0/services/sms")
@@ -413,7 +409,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
                 codeLifetime = body.smsInput.codeLifetime;
                 urlWithLinkCode = body.smsInput.urlWithLinkCode;
                 userInputCode = body.smsInput.userInputCode;
-                isFirstFactor = body.smsInput.isFirstFactor;
                 return {};
             });
 
@@ -435,7 +430,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         assert.notStrictEqual(userInputCode, undefined);
         assert.notStrictEqual(codeLifetime, undefined);
         assert(codeLifetime > 0);
-        assert(isFirstFactor);
         assert.strictEqual(result.status, 500);
         assert(message === "Request failed with status code 500");
     });
@@ -559,7 +553,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         let codeLifetime = undefined;
         let urlWithLinkCode = undefined;
         let userInputCode = undefined;
-        let isFirstFactor = undefined;
 
         nock("https://api.supertokens.com")
             .post("/0/services/sms")
@@ -569,7 +562,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
                 codeLifetime = body.smsInput.codeLifetime;
                 urlWithLinkCode = body.smsInput.urlWithLinkCode;
                 userInputCode = body.smsInput.userInputCode;
-                isFirstFactor = body.smsInput.isFirstFactor;
                 return {};
             });
 
@@ -591,7 +583,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         assert.notStrictEqual(userInputCode, undefined);
         assert.notStrictEqual(codeLifetime, undefined);
         assert(codeLifetime > 0);
-        assert(isFirstFactor);
         assert.strictEqual(result.body.status, "OK");
     });
 
@@ -632,7 +623,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         let urlWithLinkCode = undefined;
         let userInputCode = undefined;
         let loginCalled = false;
-        let isFirstFactor = undefined;
         let apiKey = "randomKey";
 
         nock("https://api.supertokens.com")
@@ -661,7 +651,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
                 codeLifetime = body.smsInput.codeLifetime;
                 urlWithLinkCode = body.smsInput.urlWithLinkCode;
                 userInputCode = body.smsInput.userInputCode;
-                isFirstFactor = body.smsInput.isFirstFactor;
                 apiKey = body.apiKey;
                 return {};
             });
@@ -683,7 +672,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         assert.notStrictEqual(userInputCode, undefined);
         assert.notStrictEqual(codeLifetime, undefined);
         assert.strictEqual(apiKey, undefined);
-        assert(isFirstFactor);
         assert(codeLifetime > 0);
     });
 
@@ -1062,7 +1050,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         let codeLifetime = undefined;
         let urlWithLinkCode = undefined;
         let userInputCode = undefined;
-        let isFirstFactor = undefined;
         let loginCalled = false;
 
         nock("https://api.supertokens.com")
@@ -1097,7 +1084,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
                 codeLifetime = body.smsInput.codeLifetime;
                 urlWithLinkCode = body.smsInput.urlWithLinkCode;
                 userInputCode = body.smsInput.userInputCode;
-                isFirstFactor = body.smsInput.isFirstFactor;
                 return {};
             });
 
@@ -1118,7 +1104,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         assert.notStrictEqual(userInputCode, undefined);
         assert.notStrictEqual(codeLifetime, undefined);
         assert(codeLifetime > 0);
-        assert(isFirstFactor);
         assert.strictEqual(result.status, 500);
         assert(message === "Request failed with status code 500");
     });
@@ -1132,7 +1117,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         let supertokensAPICalled = false;
         let apiKey = undefined;
         let type = undefined;
-        let isFirstFactor = undefined;
         let loginCalled = false;
         STExpress.init({
             supertokens: {
@@ -1201,7 +1185,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
                 phoneNumber = body.smsInput.phoneNumber;
                 codeLifetime = body.smsInput.codeLifetime;
                 type = body.smsInput.type;
-                isFirstFactor = body.smsInput.isFirstFactor;
                 return {};
             });
 
@@ -1218,7 +1201,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
 
         assert.strictEqual(phoneNumber, "+919909909998");
         assert.strictEqual(type, "PASSWORDLESS_LOGIN");
-        assert(isFirstFactor);
         assert(outerOverrideCalled);
         assert(supertokensAPICalled);
         assert.notStrictEqual(userInputCode, undefined);
@@ -1263,7 +1245,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         let codeLifetime = undefined;
         let urlWithLinkCode = undefined;
         let userInputCode = undefined;
-        let isFirstFactor = undefined;
         let loginCalled = false;
 
         nock("https://api.supertokens.com")
@@ -1292,7 +1273,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
                 codeLifetime = body.smsInput.codeLifetime;
                 urlWithLinkCode = body.smsInput.urlWithLinkCode;
                 userInputCode = body.smsInput.userInputCode;
-                isFirstFactor = body.smsInput.isFirstFactor;
                 return {};
             });
 
@@ -1313,7 +1293,6 @@ describe(`smsDelivery: ${printPath("[test/thirdpartypasswordless/smsDelivery.tes
         assert.notStrictEqual(userInputCode, undefined);
         assert.notStrictEqual(codeLifetime, undefined);
         assert(codeLifetime > 0);
-        assert(isFirstFactor);
         assert.strictEqual(result.body.status, "OK");
     });
 });